+++ jest.config.json
@@ -14,5 +14,5 @@
     "moduleNameMapper": {
         "src/(.*)": "<rootDir>/src/$1"
     },
-    "moduleDirectories": ["src", ""]
+    "moduleDirectories": ["src", "node_modules"]
 }
