+++ src/learningTypescript/part01_basic_types/part2_type_assertion/ErrorServerResult.ts
@@ -1,0 +1,3 @@
+export class ErrorServerResult {
+    constructor(public httpCode: number, public message: string) {}
+}
