+++ src/learningTypescript/part01_basic_types/part2_type_assertion/SuccessServerResult.ts
@@ -1,0 +1,6 @@
+export class SuccessServerResult {
+    constructor(
+        public httpCode: number,
+        public resultObject: Record<string, unknown>, // originally Object, @typescript-eslint/ban-types
+    ) {}
+}
