+++ src/learningTypescript/part02_classes_interfaces/part3_implement-interface-methods/Circle.ts
@@ -1,0 +1,13 @@
+import { IGeometricFigure } from '.';
+
+export class Circle implements IGeometricFigure {
+    constructor(readonly radius: number) {}
+
+    area(): number {
+        return Math.PI * this.radius ** 2;
+    }
+
+    perimeter(): number {
+        return 2 * Math.PI * this.radius;
+    }
+}
