+++ src/learningTypescript/part02_classes_interfaces/part3_implement-interface-methods/Square.ts
@@ -1,0 +1,13 @@
+import { IGeometricFigure } from '.';
+
+export class Square implements IGeometricFigure {
+    constructor(readonly side: number) {}
+
+    area(): number {
+        return this.side ** 2;
+    }
+
+    perimeter(): number {
+        return this.side * 4;
+    }
+}
