+++ src/learningTypescript/part02_classes_interfaces/part3_implement-interface-methods/index.ts
@@ -25,32 +25,7 @@ export class Circle implements IGeometricFigure {
 }
 */
 
-// eslint-disable-next-line max-classes-per-file
-interface IGeometricFigure {
+export interface IGeometricFigure {
     area: () => number;
     perimeter: () => number;
-}
-
-export class Square implements IGeometricFigure {
-    constructor(readonly side: number) {}
-
-    area(): number {
-        return this.side ** 2;
-    }
-
-    perimeter(): number {
-        return this.side * 4;
-    }
-}
-
-export class Circle implements IGeometricFigure {
-    constructor(readonly radius: number) {}
-
-    area(): number {
-        return Math.PI * this.radius ** 2;
-    }
-
-    perimeter(): number {
-        return 2 * Math.PI * this.radius;
-    }
 }
