+++ src/learningTypescript/part02_classes_interfaces/part4_abstract-classes/Boa.ts
@@ -1,0 +1,7 @@
+import { Animal } from '.';
+
+export class Boa extends Animal {
+    public constructor() {
+        super(1);
+    }
+}
