+++ src/learningTypescript/part02_classes_interfaces/part4_abstract-classes/Monkey.ts
@@ -1,0 +1,7 @@
+import { Animal } from '.';
+
+export class Monkey extends Animal {
+    public constructor() {
+        super(5);
+    }
+}
