+++ src/learningTypescript/part02_classes_interfaces/part4_abstract-classes/Parrot.ts
@@ -1,0 +1,7 @@
+import { Animal } from '.';
+
+export class Parrot extends Animal {
+    public constructor() {
+        super(38);
+    }
+}
