+++ src/learningTypescript/part02_classes_interfaces/part4_abstract-classes/index.test.ts
@@ -1,4 +1,6 @@
-import { Boa, Monkey, Parrot } from '.';
+import { Boa } from './Boa';
+import { Monkey } from './Monkey';
+import { Parrot } from './Parrot';
 
 describe(Boa.name, () => {
     it('should be measured in Monkeys correctly', () => {
