+++ src/learningTypescript/part02_classes_interfaces/part5_inheritance/Cat.ts
@@ -1,0 +1,15 @@
+import { Animal } from '.';
+
+export class Cat extends Animal {
+    constructor(
+        public name: string,
+        public age: number,
+        public status: string,
+    ) {
+        super(name, age, 4, 'cat', status);
+    }
+
+    introduce() {
+        return `${super.introduce()}  Meow meow!`;
+    }
+}
