+++ src/learningTypescript/part02_classes_interfaces/part5_inheritance/Shark.ts
@@ -1,0 +1,11 @@
+import { Animal } from '.';
+
+export class Shark extends Animal {
+    constructor(
+        public name: string,
+        public age: number,
+        public status: string,
+    ) {
+        super(name, age, 0, 'shark', status);
+    }
+}
