+++ src/learningTypescript/part02_classes_interfaces/part5_inheritance/index.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-classes-per-file */
 /*
 https://www.codewars.com/kata/59138d93eefeaecdbe000031
 
@@ -52,45 +51,5 @@ export class Animal implements IAnimal {
 
     introduce() {
         return `Hello, my name is ${this.name} and I am ${this.age} years old.`;
-    }
-}
-
-export class Shark extends Animal {
-    constructor(
-        public name: string,
-        public age: number,
-        public status: string,
-    ) {
-        super(name, age, 0, 'shark', status);
-    }
-}
-
-export class Cat extends Animal {
-    constructor(
-        public name: string,
-        public age: number,
-        public status: string,
-    ) {
-        super(name, age, 4, 'cat', status);
-    }
-
-    introduce() {
-        return `${super.introduce()}  Meow meow!`;
-    }
-}
-
-export class Dog extends Animal {
-    constructor(
-        public name: string,
-        public age: number,
-        public status: string,
-        public master: string,
-    ) {
-        super(name, age, 4, 'dog', status);
-        this.master = master;
-    }
-
-    greetMaster() {
-        return `Hello ${this.master}`;
     }
 }
