+++ src/learningTypescript/part02_classes_interfaces/part6_getters/Cube.ts
@@ -1,0 +1,7 @@
+import { Cuboid } from '.';
+
+export class Cube extends Cuboid {
+    constructor(length: number) {
+        super(length, length, length);
+    }
+}
