+++ src/learningTypescript/part02_classes_interfaces/part6_getters/index.test.ts
@@ -1,4 +1,5 @@
-import { Cube, Cuboid } from '.';
+import { Cuboid } from '.';
+import { Cube } from './Cube';
 
 describe(`class ${Cuboid.name}`, () => {
     it('should initialize properly and have working getters', () => {
