+++ src/learningTypescript/part02_classes_interfaces/part6_getters/index.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-classes-per-file */
 /*
 https://www.codewars.com/kata/599d636e607719edf600010a
 
@@ -42,11 +41,5 @@ export class Cuboid {
 
     public get volume() {
         return this.length * this.width * this.height;
-    }
-}
-
-export class Cube extends Cuboid {
-    constructor(length: number) {
-        super(length, length, length);
     }
 }
