+++ src/learningTypescript/part04_mixins/part11_mixins/index.test.ts
@@ -1,7 +1,8 @@
-/* eslint-disable @typescript-eslint/ban-ts-comment */
-/* eslint-disable max-classes-per-file */
 import { Serializable } from '.';
+import { Person } from './Person';
+import { Robot } from './Robot';
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 function applyMixins(derivedCtor: any, baseCtors: any[]) {
     baseCtors.forEach((baseCtor) => {
         Object.getOwnPropertyNames(baseCtor.prototype).forEach((name) => {
@@ -11,22 +12,9 @@ function applyMixins(derivedCtor: any, baseCtors: any[]) {
     });
 }
 
-class Person implements Serializable {
-    constructor(public firstName?: string, public lastName?: string) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    // @ts-ignore strictPropertyInitialization
-    serialize: () => string;
-
-    // @ts-ignore strictPropertyInitialization
-    deserialize: (source: string) => void;
-}
-
 applyMixins(Person, [Serializable]);
 
-describe('Person', () => {
+describe(`${Person.name}`, () => {
     it('should serialize/deserialize correctly', () => {
         const sourcePerson = new Person('John', 'Galt');
         const serializedPerson = sourcePerson.serialize();
@@ -37,24 +25,9 @@ describe('Person', () => {
     });
 });
 
-class Robot implements Serializable {
-    id: string;
-
-    constructor(public name?: string) {
-        this.id = Math.random().toString(36).substring(7);
-        this.name = name;
-    }
-
-    // @ts-ignore strictPropertyInitialization
-    serialize: () => string;
-
-    // @ts-ignore strictPropertyInitialization
-    deserialize: (source: string) => void;
-}
-
 applyMixins(Robot, [Serializable]);
 
-describe('Robot', () => {
+describe(`${Robot.name}`, () => {
     it('should serialize/deserialize correctly', () => {
         const sourceRobot = new Robot('Bender');
         const serializedRobot = sourceRobot.serialize();
